requirements.txt
beautifulsoup4==4.9.3
bs4==0.0.1
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
idna==2.10
Pillow==8.2.0
pyzbar==0.1.8
qrcode==6.1
requests==2.25.1
six==1.16.0
soupsieve==2.2.1
urllib3==1.26.5
